pyproject.toml
[project]
name = "coscientist"
version = "0.1.0"
description = "AI assistant for chemical tasks"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pubchempy (==1.0.4)",
    "py3dmol (==2.4.2)",
    "rdkit (==2024.9.6)",
    "langchain-experimental (==0.3.4)",
    "gdown (==5.2.0)",
    "torch (==2.6.0)",
    "torchvision (==0.21.0)",
    "duckduckgo-search (==8.0.3)",
    "smolagents (==1.18.0)",
    "deepeval (==2.3.3)",
    "docling (==2.28.4)",
    "docling-core (==2.26.0)",
    "python-dotenv (==1.0.1)",
    "langchain (==0.3.22)",
    "langchain-core (==0.3.50)",
    "pandas (==2.2.3)",
    "pillow (==10.4.0)",
    "pydantic (==2.9.2)",
    "open-clip-torch (==2.32.0)",
    "sentence-transformers (==4.1.0)",
    "chromadb (==1.0.6)",
    "streamlit (==1.45.1)",
    "langchain-gigachat (==0.3.3)",
    "langchain-ollama (==0.3.0)",
    "langchain-openai (==0.3.3)",
    "marker-pdf (==1.6.2)",
    "beautifulsoup4 (==4.13.3)",
    "streamlit-pills (==0.3.0)",
    "streamlit_extras (==0.7.5)",
    "pyperclip (==1.9.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
